Add test unit to breaktrough
I added in the test directory the breaktrough_testing.py file containing
a test unit for the breaktrough game. It simply challenge most of the
functions that are defined inside the game file. The ones requiring user
inputs aren't tested. After writing the tests, some were failing, due to
errors in the original game code. These mistakes are now things of the
past.

Writing tests helps detecting potential errors that crept into a source
code. I wanted to experiment what it might be like to write tests units
and the benefits that can come with it.

The unittest module is used to create the unit test. The
BreaktroughTests class contains methods that test most of the functions
used in the game in various cases.

A .gitignore file was added, currently mainly to ignore the __pycache__
directories.