Add AI and its tests. Remove facing squares bug
A random playing AI is now implemented. Tested, it unveiled a mistake in
the pawn_facing_squares function that, when computing squares in front a
pawn that is on the opponent line, was trying to access a too high y
index value, outputing a dirty error message.
Pylint's configuration file now states that docstrings are optional, as
it was logging a lot of warning messages while scanning the breaktrough
unit test.
Also, fine tuned console parameters treament when script is executed,
and indicates user what parameters are required to launch it.