Make unavailable facing takes in Breaktrough
Re-reading the subject, I noticed that, in the Breaktrough game, a pawn
cannot take an enemy pawn that is in front of it. Only diagonaly
adjacent pawns can be taken. This slip is now overwritten with
conditions that match the rules. Tests have been changed to make them
check if this rule is respected.