Fix language/caps inconsistencies
Mainly caps words that should have been earlier and also some grammar,
like the impersonal/personal "there/here". Also add function annotations
to the functions that were missing some.
In First Attack, rename empty_squares_with_values to flush_values to
express clearer what the function does, and does not anymore return a
flattened board but delegate it to the user with flatten function,
being more coherent with the rest of the code.